requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tkinterdnd2>=0.3.0
Pillow>=10.0.0 